delete clearanceitem, rename search.py
