+++ acslib/ccure/__init__.py
@@ -2,4 +2,4 @@ from acslib.base import ACSRequestResponse
 from acslib.base.connection import ACSRequestData, ACSRequestMethod
 from acslib.ccure.base import CcureACS
 from acslib.ccure.crud import CcureAPI, CCurePersonnel, CCureClearance, CCureCredential
-from acslib.ccure.search import ClearanceFilter, PersonnelFilter, CredentialFilter
+from acslib.ccure.search_filtering import ClearanceFilter, PersonnelFilter, CredentialFilter
